pyproject.toml
[project]
name = "raft"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "markitdown[pdf]>=0.0.1a0",
    "requests>=2.31.0",
    "psycopg[binary]>=3.1.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]
